run VIM prettier on all files

+++ contracts/artifacts/MappingEx.sol
@@ -1,13 +1,13 @@
 pragma solidity ^0.5.13;
 
 contract SimpleMappingExample {
-    mapping(uint => bool) public myMapping;
+    mapping(uint256 => bool) public myMapping;
     mapping(address => bool) public addressMapping;
-    
-    function setValue(uint _index) public {
+
+    function setValue(uint256 _index) public {
         myMapping[_index] = true;
     }
-    
+
     function setMyAddressToTrue() public {
         addressMapping[msg.sender] = true;
     }
